fix bugs

+++ lib/utils/linemod/opengl_renderer.py
@@ -71,26 +71,26 @@ class OpenGLRenderer(object):
 
         # Prepare data structures
         model = {}
-        model['pts'] = np.zeros((n_pts, 3), np.float)
+        model['pts'] = np.zeros((n_pts, 3), float)
         if n_faces > 0:
-            model['faces'] = np.zeros((n_faces, face_n_corners), np.float)
+            model['faces'] = np.zeros((n_faces, face_n_corners), float)
 
         pt_props_names = [p[0] for p in pt_props]
         is_normal = False
         if {'nx', 'ny', 'nz'}.issubset(set(pt_props_names)):
             is_normal = True
-            model['normals'] = np.zeros((n_pts, 3), np.float)
+            model['normals'] = np.zeros((n_pts, 3), float)
 
         is_color = False
-        model['colors'] = np.zeros((n_pts, 3), np.float)
+        model['colors'] = np.zeros((n_pts, 3), float)
         if {'red', 'green', 'blue'}.issubset(set(pt_props_names)):
             is_color = True
-            model['colors'] = np.zeros((n_pts, 3), np.float)
+            model['colors'] = np.zeros((n_pts, 3), float)
 
         is_texture = False
         if {'texture_u', 'texture_v'}.issubset(set(pt_props_names)):
             is_texture = True
-            model['texture_uv'] = np.zeros((n_pts, 2), np.float)
+            model['texture_uv'] = np.zeros((n_pts, 2), float)
 
         formats = { # For binary format
             'float': ('f', 4),
